handling malformed message, a message that does not contain a colon

+++ CSE546WebtierImage/imagequeue.py
@@ -120,6 +120,8 @@ def receiverThreadFunction() :
     while(True) :
         for message in ResponseQueue.receive_messages(WaitTimeSeconds=20):
             message_content = message.body
+            if(message_content.find(":") == -1)
+                continue
             image_filename, image_result = parse_message(message_content)
             setResultToReceivedMessage(image_filename, image_result)
             message.delete()
